Go over the default.cfg's of all the example games. Make it so that Skill_GetValue() requires a default value set.

+++ src/gs-entbase/server/basemonster.h
@@ -98,7 +98,7 @@ class CBaseMonster:CBaseEntity
 	int m_iFlags;
 	vector base_mins;
 	vector base_maxs;
-	int base_health;
+	float base_health;
 
 	float m_flChaseSpeed;
